MOD-501: Setup Swagger API documentation for moderation endpoints
