+++ package.json
@@ -5,7 +5,8 @@
 	"main": "dist/server.js",
 	"scripts": {
 		"start": "ts-node src/server.ts",
-		"build": "tsc"
+		"build": "tsc",
+		"generate-docs": "ts-node scripts/generate-openapi.ts && npx @redocly/cli build-docs openapi.json --output api-docs/index.html"
 	},
 	"dependencies": {
 		"dotenv": "^16.4.7",
@@ -15,6 +16,7 @@
 		"swagger-ui-express": "^5.0.1"
 	},
 	"devDependencies": {
+		"@redocly/cli": "^1.33.1",
 		"@types/express": "^4.17.21",
 		"@types/node": "^20.12.13",
 		"@types/swagger-jsdoc": "^6.0.4",
