+++ src/app.ts
@@ -1,6 +1,6 @@
 import express, { Express, Request, Response } from "express";
 import moderationRoutes from "./api/v1/routes/moderationRoutes";
-import { setupSwagger } from "./config/swagger";
+import setupSwagger from "./config/swagger";
 
 const app: Express = express();
 app.use(express.json());
